code clean & format client

+++ hugegraph-client/src/main/java/org/apache/hugegraph/api/traverser/NeighborRankAPI.java
@@ -140,7 +140,7 @@ public class NeighborRankAPI extends TraversersAPI {
                 this.direction = null;
                 this.labels = new ArrayList<>();
                 this.degree = Traverser.DEFAULT_MAX_DEGREE;
-                this.top = (int) Traverser.DEFAULT_PATHS_LIMIT;
+                this.top = Traverser.DEFAULT_PATHS_LIMIT;
             }
 
             @Override
